Fix a bug in the J2 plasticity code.
The `dLbar` Tensor was being incorrectly initialized. It was supposed to capture
an external pointer for use as an ouput buffer, but it wasn't initialized to
anything.

As a side effect of this change I fixed some of the syle problems with this
file. In particular, when the type of a parameter is not explicitly required
inside of a function call it makes sense to use `auto` rather than explicitly
introducing a type parameter name.